feat: primitive fields
